Create a seperate member provider for serialize fun spec as well
